+++ Source/ProyectHades/Private/DataAssets/StartUpData/DataAsset_HeroStartUpData.cpp
@@ -5,11 +5,6 @@
 #include "AbiilitySystem/Abilities/WarriorGameplayAbility.h"
 #include "AbiilitySystem/WarriorAbilitySystemComponent.h"
 
-bool FWarriorHeroAbilitySet::IsValid() const
-{
-	return InputTag.IsValid() && AbilityToGrant;
-}
-
 void UDataAsset_HeroStartUpData::GiveToAbilitySystemComponent(UWarriorAbilitySystemComponent* InASCToGive,
 	int32 ApplyLevel)
 {
