+++ Source/ProyectHades/Private/WarriorGameplayTags.cpp
@@ -13,6 +13,9 @@ namespace  WarriorGameplayTags
 	/** Player TAGS **/
 	UE_DEFINE_GAMEPLAY_TAG(Player_Weapon_Axe, "Player.Weapon.Axe");
 	
+	UE_DEFINE_GAMEPLAY_TAG(Player_Ability_Equip_Axe, "Player.Ability.Equip.Axe");
+	UE_DEFINE_GAMEPLAY_TAG(Player_Ability_Unequip_Axe, "Player.Ability.Unequip.Axe");	
+	
 	UE_DEFINE_GAMEPLAY_TAG(Player_Events_Equip_Axe, "Player.Event.Equip.Axe");
 	UE_DEFINE_GAMEPLAY_TAG(Player_Events_Unequip_Axe, "Player.Event.Unequip.Axe");
 }
