+++ Source/ProyectHades/Public/WarriorGameplayTags.h
@@ -12,8 +12,13 @@ namespace  WarriorGameplayTags
 	PROYECTHADES_API UE_DECLARE_GAMEPLAY_TAG_EXTERN(InputTag_EquipAxe);
 	PROYECTHADES_API UE_DECLARE_GAMEPLAY_TAG_EXTERN(InputTag_UnequipAxe);
 	/** Player TAGS **/
+
+	PROYECTHADES_API UE_DECLARE_GAMEPLAY_TAG_EXTERN(Player_Ability_Equip_Axe);
+	PROYECTHADES_API UE_DECLARE_GAMEPLAY_TAG_EXTERN(Player_Ability_Unequip_Axe);
+	
 	PROYECTHADES_API UE_DECLARE_GAMEPLAY_TAG_EXTERN(Player_Weapon_Axe);
 
 	PROYECTHADES_API UE_DECLARE_GAMEPLAY_TAG_EXTERN(Player_Events_Equip_Axe);
 	PROYECTHADES_API UE_DECLARE_GAMEPLAY_TAG_EXTERN(Player_Events_Unequip_Axe);
+	
 }
