+++ Source/ProyectHades/Public/WarriorTypes/WarriorStructTypes.h
@@ -1,12 +1,29 @@
 // KasaiRaito Studios All Rights Reserved
 
 #pragma once
-#include "WarriorStructTypes.generated.h"\
+#include "GameplayTagContainer.h"
+#include "WarriorStructTypes.generated.h"
 
 class UWarriorHeroLinkedAnimLayer;
+class UWarriorGameplayAbility;
+class UInputMappingContext;
 /**
  * 
  */
+USTRUCT(BlueprintType)
+struct FWarriorHeroAbilitySet
+{
+	GENERATED_BODY()
+
+	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly, meta = (Categories = "InputTag"))
+	FGameplayTag InputTag;
+
+	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly)
+	TSubclassOf<UWarriorGameplayAbility> AbilityToGrant;
+
+	bool IsValid() const;
+};
+
 USTRUCT(BlueprintType)
 struct FWarriorHeroWeaponData
 {
@@ -14,4 +31,10 @@ struct FWarriorHeroWeaponData
 
 	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly)
 	TSubclassOf<UWarriorHeroLinkedAnimLayer> WeaponAnimLayerToLink;
+	
+	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly)
+	UInputMappingContext* WeaponInputMappingContext;
+
+	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly, meta = (TitleProperty = "InputTag"))
+	TArray<FWarriorHeroAbilitySet> DefaultWeaponAbilities;
 };
